+++ core-bundle/src/Resources/contao/dca/tl_content.php
@@ -1181,7 +1181,7 @@ class tl_content extends Backend
 		// Limit the element's height
 		if (!Config::get('doNotCollapse'))
 		{
-			$class .=  ' h64';
+			$class .=  ' h40';
 		}
 
 		$objModel = new ContentModel();